rag/requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
chromadb
langchain-community
pypdf
python-dotenv